+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 
-## [0.5.3](https://github.com/Blobfolio/dactyl/releases/tag/v0.5.3) - TBD
+## [0.6.0](https://github.com/Blobfolio/dactyl/releases/tag/v0.6.0) - 2023-10-15
 
 ### New